Fixed endpoints and attempted to implement elastic search

+++ api/poems.js
@@ -1,8 +1,19 @@
 const router = require('express').Router();
 const fs = require('fs');
-
 const { validateAgainstSchema } = require('../lib/validation');
 const { generateAuthToken, requireAuthentication, optionalAuthentication } = require('../lib/auth');
+const mongoose = require('mongoose');
+const mongoosastic = require('mongoosastic');
+
+
+var m_PoemSchema = new mongoose.Schema({
+    
+    author: {type: String, es_indexed: true },
+    uploader: {type: String, es_indexed: true},
+    text: {type: String, es_indexed: true},
+    category: {type: String, es_indexed: true},
+    date: {type: String, es_indexed: true }
+});
 
 
 module.exports = router;
@@ -19,6 +30,8 @@ const {
     deletePoemById,
 
 } =  require('../models/poem');
+const { getAdminStatus } = require('../models/user');
+const { getDbReference } = require('../lib/mongo');
 
 //Post a new poem
 router.post('/', requireAuthentication, async(req, res, next) => {
@@ -143,12 +156,13 @@ router.put('/:id', requireAuthentication, async(req, res, next) => {
         try{
 
             var results = await updatePoemById(req.params.id, req.body)
-    
-            console.log(results)
+
+            if(results){
+
+                res.status(200).send(results);
+            }
+
             
-            res.status(200).send(results);
-    
-    
         }catch(error){
     
             console.log(error)
@@ -179,12 +193,19 @@ router.delete('/:id', requireAuthentication, async(req, res, next) => {
 
     try{
 
-        var results = await deletePoemById(req.params.id)
+        const isAdmin = await getAdminStatus(req.username);
+        
+        if(isAdmin){
+
+            var results = await deletePoemById(req.params.id)
+
+        }
 
         if(req.username === results.uploader){
 
+            var results = await deletePoemById(req.params.id)
             
-        res.status(200).send(results);
+            res.status(200).send(results);
 
         } else {
 
@@ -210,10 +231,67 @@ router.delete('/:id', requireAuthentication, async(req, res, next) => {
 
 
 //Get Via Elastic Search(idk how to do this)
-router.get('/search/:searchToken', requireAuthentication, async(req, res, next) => {
+router.get('/search/:searchString', async(req, res, next) => {
+
+    m_PoemSchema.plugin(mongoosastic)
+
+    var Poem = mongoose.model('Poem', m_PoemSchema, 'poems')
+
+    var stream = Poem.synchronize();
+    
+    stream.on('error', function(err){
+
+        console.log("Error while syncing" + err)
+
+    });
 
 
+    const db = getDbReference();
 
+    const collection = db.collection('poems');
 
+    var types = ['poems']
 
+    var fields = ['text', 'author', 'category']
+
+    var filter = {}
+    var sort = "_id";
+
+    Poem.search({
+        bool: {
+            must: [{
+              multi_match: {
+                query: req.params.searchString,
+                type: "phrase",
+                fields: fields,
+              },
+            }],
+            should: [{
+              multi_match: {
+                query: req.params.searchString,
+                fields: fields,
+                operator: "and",
+                boost: 10,
+                minimum_should_match: "100%"
+              },
+            }],
+            filter: [{
+              term: filter
+            }]
+           }
+         }, {
+           index: collection,
+           type: types,
+           from: null,
+           size: 10,
+           track_scores: false,
+           sort: sort
+          },
+          function (err, results) {
+            if (err) {
+               console.log("Error in searchController: ", err);
+             } else {
+               res.json({result: results.hits.hits})
+             }
+    });
 });
